override link as a func param
